api/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.0
graphql-core==3.2.3
h11==0.14.0
idna==3.4
pydantic==1.10.7
python-dateutil==2.8.2
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
strawberry-graphql==0.168.1
typing_extensions==4.5.0
uvicorn==0.21.1
